feat!: Report catalog version as a label
In order to accommodate environments where catalog versions are not numbers,
`puppet_last_catalog_version` now reports a `version` label and always has a
value of `1`.

This is a breaking change because it makes alerting on the staleness of the
catalog version impossible, but in practice that alert was never useful. If
you had the `StalePuppetCatalog` sample rule in your prometheus configuration,
you should remove it before upgrading the exporter, in order to avoid false
alerts caused by the value now being `1`.

Closes #60

Co-authored-by: Andrew Seales <andrew.seales@ed.ac.uk>
Signed-off-by: Erik Swanson <erik@retailnext.net>