expose puppet file locations as commandline arguments
the location of the run report changed with puppet7. the debian
package patches puppet to look for the configuration file under
/etc/puppet and the report is stored under /var/cache/puppet.
other OS packages might use different paths as well. this change
exposes the already present file locations to be adjusted via
commandline arguments. the default values remain the same.

Signed-off-by: Gordon Bleux <33967640+UiP9AV6Y@users.noreply.github.com>